+++ src/components/core.ts
@@ -185,6 +185,7 @@ export default class Core {
       a: true,
     } as SanitizerConfig;
 
+
     this.config.hideToolbar = this.config.hideToolbar ? this.config.hideToolbar : false;
     this.config.tools = this.config.tools || {};
     this.config.i18n = this.config.i18n || {};
@@ -208,7 +209,6 @@ export default class Core {
     }
 
     this.config.readOnly = this.config.readOnly as boolean || false;
-    this.config.i18n = {};
 
     /**
      * Adjust i18n
@@ -220,11 +220,7 @@ export default class Core {
     /**
      * Text direction. If not set, uses ltr
      */
-    if (config.i18n?.direction) {
-      this.config.i18n = {
-        direction: config.i18n?.direction || 'ltr',
-      };
-    }
+    this.config.i18n.direction = config.i18n?.direction || 'ltr';
   }
 
   /**
@@ -292,8 +288,6 @@ export default class Core {
       'InlineToolbar',
       'BlockManager',
       'Paste',
-      'DragNDrop',
-      'ModificationsObserver',
       'BlockSelection',
       'RectangleSelection',
       'CrossBlockSelection',
@@ -337,7 +331,7 @@ export default class Core {
       /**
        * If module has non-default exports, passed object contains them all and default export as 'default' property
        */
-      const Module = typeof module === 'function' ? module : module.default;
+      const Module = _.isFunction(module) ? module : module.default;
 
       try {
         /**