+++ src/components/modules/toolbar/index.ts
@@ -141,9 +141,6 @@ export default class Toolbar extends Module<ToolbarNodes> {
 
   /**
    * Module preparation method
-   * Steps:
-   *  - Make Toolbar dependent components like BlockSettings, Toolbox and so on
-   *  - Make itself and append dependent nodes to itself
    */
   public async prepare(): Promise<void> {
     /**
@@ -369,6 +366,11 @@ export default class Toolbar extends Module<ToolbarNodes> {
    *
    * Toolbar contains BlockSettings and Toolbox.
    * Thats why at first we draw its components and then Toolbar itself
+   *
+   * Steps:
+   *  - Make Toolbar dependent components like BlockSettings, Toolbox and so on
+   *  - Make itself and append dependent nodes to itself
+   *
    */
   private drawUI(): void {
     /**