+++ types/api/blocks.d.ts
@@ -15,11 +15,10 @@ export interface Blocks {
    * Render passed data
    *
    * @param {OutputData} data - saved Block data
-   * @param {boolean} readOnly - the flag that should be used to render a block in the read-only mode
    *
    * @returns {Promise<void>}
    */
-  render(data: OutputData, readOnly: boolean): Promise<void>;
+  render(data: OutputData): Promise<void>;
 
   /**
    * Render passed HTML string